requirements.txt
bottle==0.12.18
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-WTF==0.14.3
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.16.0
MarkupSafe==1.1.1
numpy==1.19.2
pandas==1.1.2
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
scikit-learn==0.23.2
scipy==1.5.2
six==1.15.0
spotipy==2.16.0
threadpoolctl==2.1.0
urllib3==1.25.10
Werkzeug==1.0.1
WTForms==2.3.3
